std-toast/show-toast.http.html

<!DOCTYPE html>
<meta charset="utf-8">
<title>Toast: showToast tests</title>

<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>

<body>
</body>

<script type="module">
import { showToast, StdToastElement } from 'std:toast';
import { testShowToast, assertToastNotShown, assertToastShown } from './resources/helpers.js';

testShowToast((toast) => {
    assert_true(toast != null);
    assert_true(toast instanceof StdToastElement);
}, 'showToast creates and returns a toast');

testShowToast((toast) => {
    assert_true(document.querySelector('std-toast') != null);
}, 'showToast puts the toast in the DOM');

testShowToast((toast) => {
    assertToastShown(toast);
}, 'showToast displays the toast by default');

testShowToast((toast) => {
    toast.hide();

    assertToastNotShown(toast);
}, 'hiding showToast immediately does not display it');

testShowToast((toast) => {
    assert_true(toast.hasAttribute('open'));
}, 'the open attribute is present upon creation');

testShowToast((toast) => {
    toast.show();

    assertToastShown(toast);
}, 'calling show after showToast does nothing');

test(() => {
    let toast1, toast2;
    try {
        toast1 = showToast('message1');
        toast2 = showToast('message2');

        assert_not_equals(toast1, toast2);
    }
    finally {
        toast1.remove();
        toast2.remove();
    }
}, 'calling showToast multiple times creates multiple different toasts');

testShowToast((toast) => {
    assert_equals(toast.textContent, 'message');
}, 'showToast created toast has `message` as its children');

</script>
